requirements.txt
bcrypt==3.2.2
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
Flask==2.2.2
flask-grpc==0.1.1
Flask-Session==0.4.0
Flask-WTF==1.0.1
grpcio==1.47.0
grpcio-tools==1.47.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
protobuf==3.20.1
pycparser==2.21
pymongo==4.2.0
six==1.16.0
Werkzeug==2.2.2
WTForms==3.0.1
zipp==3.8.1
